Fix comparison between signed and unsigned int.

+++ src/io/config.cpp
@@ -184,7 +184,7 @@ void Config::GetAucMuWeights(const std::unordered_map<std::string, std::string>&
                      num_class, oneline_weights.size());
         }
         std::vector<double> curr_line_weights;
-        for (size_t j = 0; j < num_class; ++j) {
+        for (size_t j = 0; j < static_cast<size_t>(num_class); ++j) {
           curr_line_weights.push_back(oneline_weights[j]);
           if (oneline_weights[j] < 0) {
             Log::Fatal("Auc-mu weights matrix must contain only non-negative values. Found negative value at position [%d, %d]",
@@ -200,13 +200,13 @@ void Config::GetAucMuWeights(const std::unordered_map<std::string, std::string>&
       }
     };
   reader.ReadAllAndProcessParallel(process_fun);
-  if (auc_mu_weights.size() != num_class) {
+  if (auc_mu_weights.size() != static_cast<size_t>(num_class)) {
     Log::Fatal("Auc-mu matrix must have %d rows but found %d", num_class, auc_mu_weights.size());
   }
   } else {
     // equal weights for all classes
     auc_mu_weights = std::vector<std::vector<double>> (num_class, std::vector<double>(num_class, 1));
-    for (size_t i = 0; i < num_class; ++i) {
+    for (size_t i = 0; i < static_cast<size_t>(num_class); ++i) {
       auc_mu_weights[i][i] = 0;
     }
   }
